requirements.txt
streamlit
tensorflow==2.15.0 
pillow
matplotlib